move isEmptyObject and clone into a utils file.
