--- src/empty-object.js
@@ -1,5 +1,0 @@
-export default function isEmptyObject(obj) {
-  for (let key in obj) return false
-
-  return true
-}
